Convert rake task minitest to rspec
