+++ boxt_ruby_style_guide.gemspec
@@ -31,6 +31,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency "rubocop-faker", "~> 1.0"
   spec.add_dependency "rubocop-rails", "~> 2.6"
   spec.add_dependency "rubocop-rspec", "~> 1.43"
+  spec.add_development_dependency "rails", "> 5"
   spec.add_development_dependency "bundler", "~> 2.1"
   spec.add_development_dependency "rake", "~> 13.0"
   spec.add_development_dependency "rspec-nc", "~> 0.3"
